+++ sdk/core/azure-core/inc/azure/core/case_insensitive_containers.hpp
@@ -11,6 +11,7 @@
 #include "azure/core/internal/strings.hpp"
 
 #include <map>
+#include <set>
 #include <string>
 
 namespace Azure { namespace Core {
@@ -21,5 +22,10 @@ namespace Azure { namespace Core {
    */
   using CaseInsensitiveMap
       = std::map<std::string, std::string, Internal::Strings::CaseInsensitiveComparator>;
+
+  /**
+   * @brief A type alias of `std::set<std::string>` with case-insensitive element comparison.
+   */
+  using CaseInsensitiveSet = std::set<std::string, Internal::Strings::CaseInsensitiveComparator>;
 
 }} // namespace Azure::Core