+++ sdk/core/azure-core/inc/azure/core/http/http.hpp
@@ -8,7 +8,7 @@
 
 #pragma once
 
-#include "azure/core/case_insensitive_map.hpp"
+#include "azure/core/case_insensitive_containers.hpp"
 #include "azure/core/exception.hpp"
 #include "azure/core/internal/contract.hpp"
 #include "azure/core/io/body_stream.hpp"
@@ -52,6 +52,25 @@ namespace Azure { namespace Core { namespace Http {
         CaseInsensitiveMap& headers,
         std::string const& headerName,
         std::string const& headerValue);
+
+    inline std::string FormatEncodedUrlQueryParameters(
+        std::map<std::string, std::string> const& encodedQueryParameters)
+    {
+      {
+        std::string queryStr;
+        if (!encodedQueryParameters.empty())
+        {
+          auto separ = '?';
+          for (const auto& q : encodedQueryParameters)
+          {
+            queryStr += separ + q.first + '=' + q.second;
+            separ = '&';
+          }
+        }
+
+        return queryStr;
+      }
+    }
   } // namespace Details
 
   /*********************  Exceptions  **********************/
@@ -260,6 +279,8 @@ namespace Azure { namespace Core { namespace Http {
     // this set.
     const static std::unordered_set<unsigned char> defaultNonUrlEncodeChars;
 
+    std::string GetUrlWithoutQuery(bool relative) const;
+
   public:
     /**
      * @brief Decodes \p value by transforming all escaped characters to it's non-encoded value.
@@ -423,6 +444,12 @@ namespace Azure { namespace Core { namespace Http {
     {
       return m_encodedQueryParameters;
     }
+
+    /**
+     * @brief Get Scheme, host, and path, without query parameters.
+     * @return Absolute URL without query parameters.
+     */
+    std::string GetUrlWithoutQuery() const { return GetUrlWithoutQuery(false); }
 
     /**
      * @brief Get the path and query parameters.
@@ -663,7 +690,7 @@ namespace Azure { namespace Core { namespace Http {
     // ===== Methods used to build HTTP response =====
 
     /**
-     * @brief Set an HTTP header to the #Azure::Core::Http::Request.
+     * @brief Set an HTTP header to the #RawResponse.
      *
      * @remark The \p name must contain valid header name characters (RFC 7230).
      *
@@ -675,7 +702,7 @@ namespace Azure { namespace Core { namespace Http {
     void SetHeader(std::string const& name, std::string const& value);
 
     /**
-     * @brief Set an HTTP header to the #Azure::Core::Http::Request.
+     * @brief Set an HTTP header to the #RawResponse.
      *
      * @remark The \p header must contain valid header name characters (RFC 7230).
      * @remark Header name, value and delimiter are expected to be in \p header.
@@ -687,7 +714,7 @@ namespace Azure { namespace Core { namespace Http {
     void SetHeader(std::string const& header);
 
     /**
-     * @brief Set an HTTP header to the #Azure::Core::Http::Request.
+     * @brief Set an HTTP header to the #RawResponse.
      *
      * @remark The string referenced by \p first and \p last must contain valid header name
      * characters (RFC 7230).