+++ sdk/storage/azure-storage-files-shares/src/share_file_client.cpp
@@ -106,9 +106,9 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
     }
     if (options.SmbProperties.CreatedOn.HasValue())
     {
-      protocolLayerOptions.FileCreationTime
-          = options.SmbProperties.CreatedOn.GetValue().GetRfc3339String(
-              Core::DateTime::TimeFractionFormat::AllDigits);
+      protocolLayerOptions.FileCreationTime = options.SmbProperties.CreatedOn.GetValue().ToString(
+          Azure::Core::DateTime::DateFormat::Rfc3339,
+          Core::DateTime::TimeFractionFormat::AllDigits);
     }
     else
     {
@@ -117,7 +117,8 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
     if (options.SmbProperties.LastWrittenOn.HasValue())
     {
       protocolLayerOptions.FileLastWriteTime
-          = options.SmbProperties.LastWrittenOn.GetValue().GetRfc3339String(
+          = options.SmbProperties.LastWrittenOn.GetValue().ToString(
+              Azure::Core::DateTime::DateFormat::Rfc3339,
               Core::DateTime::TimeFractionFormat::AllDigits);
     }
     else
@@ -322,7 +323,8 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
     if (options.SmbProperties.CreatedOn.HasValue())
     {
       protocolLayerOptions.FileCopyFileCreationTime
-          = options.SmbProperties.CreatedOn.GetValue().GetRfc3339String(
+          = options.SmbProperties.CreatedOn.GetValue().ToString(
+              Azure::Core::DateTime::DateFormat::Rfc3339,
               Core::DateTime::TimeFractionFormat::AllDigits);
     }
     else
@@ -332,7 +334,8 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
     if (options.SmbProperties.LastWrittenOn.HasValue())
     {
       protocolLayerOptions.FileCopyFileLastWriteTime
-          = options.SmbProperties.LastWrittenOn.GetValue().GetRfc3339String(
+          = options.SmbProperties.LastWrittenOn.GetValue().ToString(
+              Azure::Core::DateTime::DateFormat::Rfc3339,
               Core::DateTime::TimeFractionFormat::AllDigits);
     }
     else
@@ -416,7 +419,8 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
     }
     if (smbProperties.CreatedOn.HasValue())
     {
-      protocolLayerOptions.FileCreationTime = smbProperties.CreatedOn.GetValue().GetRfc3339String(
+      protocolLayerOptions.FileCreationTime = smbProperties.CreatedOn.GetValue().ToString(
+          Azure::Core::DateTime::DateFormat::Rfc3339,
           Core::DateTime::TimeFractionFormat::AllDigits);
     }
     else
@@ -425,9 +429,9 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
     }
     if (smbProperties.LastWrittenOn.HasValue())
     {
-      protocolLayerOptions.FileLastWriteTime
-          = smbProperties.LastWrittenOn.GetValue().GetRfc3339String(
-              Core::DateTime::TimeFractionFormat::AllDigits);
+      protocolLayerOptions.FileLastWriteTime = smbProperties.LastWrittenOn.GetValue().ToString(
+          Azure::Core::DateTime::DateFormat::Rfc3339,
+          Core::DateTime::TimeFractionFormat::AllDigits);
     }
     else
     {
@@ -843,9 +847,9 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
     }
     if (options.SmbProperties.CreatedOn.HasValue())
     {
-      protocolLayerOptions.FileCreationTime
-          = options.SmbProperties.CreatedOn.GetValue().GetRfc3339String(
-              Core::DateTime::TimeFractionFormat::AllDigits);
+      protocolLayerOptions.FileCreationTime = options.SmbProperties.CreatedOn.GetValue().ToString(
+          Azure::Core::DateTime::DateFormat::Rfc3339,
+          Core::DateTime::TimeFractionFormat::AllDigits);
     }
     else
     {
@@ -854,7 +858,8 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
     if (options.SmbProperties.LastWrittenOn.HasValue())
     {
       protocolLayerOptions.FileLastWriteTime
-          = options.SmbProperties.LastWrittenOn.GetValue().GetRfc3339String(
+          = options.SmbProperties.LastWrittenOn.GetValue().ToString(
+              Azure::Core::DateTime::DateFormat::Rfc3339,
               Core::DateTime::TimeFractionFormat::AllDigits);
     }
     else
@@ -948,9 +953,9 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
     }
     if (options.SmbProperties.CreatedOn.HasValue())
     {
-      protocolLayerOptions.FileCreationTime
-          = options.SmbProperties.CreatedOn.GetValue().GetRfc3339String(
-              Core::DateTime::TimeFractionFormat::AllDigits);
+      protocolLayerOptions.FileCreationTime = options.SmbProperties.CreatedOn.GetValue().ToString(
+          Azure::Core::DateTime::DateFormat::Rfc3339,
+          Core::DateTime::TimeFractionFormat::AllDigits);
     }
     else
     {
@@ -959,7 +964,8 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
     if (options.SmbProperties.LastWrittenOn.HasValue())
     {
       protocolLayerOptions.FileLastWriteTime
-          = options.SmbProperties.LastWrittenOn.GetValue().GetRfc3339String(
+          = options.SmbProperties.LastWrittenOn.GetValue().ToString(
+              Azure::Core::DateTime::DateFormat::Rfc3339,
               Core::DateTime::TimeFractionFormat::AllDigits);
     }
     else