+++ sdk/core/azure-core/CHANGELOG.md
@@ -29,6 +29,11 @@
   - `HttpPolicy::Send(Request& request, NextHttpPolicy policy, Azure::Core::Context const& context)`
   - `Operation<T>::PollUntilDone(std::chrono::milliseconds period, Azure::Core::Context& context)`
   - `TokenCredential::GetToken(Http::TokenRequestOptions const& tokenRequestOptions, Azure::Core::Context const& context)`
+- Changed type of `Azure::Core::Http::RetryOptions::StatusCodes` from `std::vector` to `std::set`.
+- Renamed `Azure::Core::Http::LoggingPolicy` to `LogPolicy`.
+- Introduced `Azure::Core::Http::LogOptions`, a mandatory parameter for `LogPolicy` construction. Entities that are not specified in the allow lists are hidden in the log.
+- Moved `Azure::Core::Logging` namespace entities to `Azure::Core::Logger` class.
+- Removed `Azure::Core::DateTime::GetRfc3339String()`: `Azure::Core::DateTime::ToString()` was extended to provide the same functionality.
 
 ### Bug Fixes